+++ packages/subagents/CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## [Unreleased]
+
+### Fixed
+
+- Capped subagent fanout spawned from workflow stages to a single child level with a workflow-specific nested-subagent error.
+
 ## [0.24.3] - 2026-05-14
 
 ### Added
@@ -13,6 +19,7 @@
 
 ### Fixed
 - Fixed builtin subagent skill resolution from project cwd ([#1087](https://github.com/flora131/atomic/issues/1087)).
+- Capped subagent fanout spawned from workflow stages to a single child level with a workflow-specific nested-subagent error.
 
 ## [0.8.18] - 2026-05-27